BotRouting Example -- Improve language style in README.md, simplify code
